Ensure 64-bit alignment for atomic structures
Operations against EWMA instances were panic'ing on i386 machines. A bug against
golang was file, and it was closed out noting: it is the caller's responsibility
to arrange for 64-bit alignment of 64-bit words accessed atomically.

This comment ensures such alignment for all instances where atomic is being used
to access their fields (ewma was the only offender).

http://code.google.com/p/go/issues/detail?id=6404